backend/requirements.txt
google-cloud-storage
google-cloud-pubsub
google-cloud-aiplatform